Moved stb .h files to deps/stb.
Added reading from memory.
